+++ server/src/models/analytic_unit_cache_model.ts
@@ -1,4 +1,4 @@
-import { AnalyticUnitId } from "./analytic_units/analytic_unit_model";
+import { AnalyticUnitId } from './analytic_units';
 import { Collection, makeDBQ } from '../services/data_service';
