+++ server/src/models/analytic_units/db.ts
@@ -1,0 +1,73 @@
+import { createAnalyticUnitFromObject } from './utils';
+import { AnalyticUnit } from './analytic_unit_model';
+import { AnalyticUnitId, FindManyQuery } from './types';
+import { Collection, makeDBQ, SortingOrder } from '../../services/data_service';
+
+import { Metric } from 'grafana-datasource-kit';
+
+
+const db = makeDBQ(Collection.ANALYTIC_UNITS);
+
+export async function findById(id: AnalyticUnitId): Promise<AnalyticUnit> {
+  let obj = await db.findOne(id);
+  if (obj === null) {
+    return null;
+  }
+  return createAnalyticUnitFromObject(obj);
+}
+
+export async function findMany(query: FindManyQuery): Promise<AnalyticUnit[]> {
+  const analyticUnits = await db.findMany(query, {
+    createdAt: SortingOrder.ASCENDING,
+    name: SortingOrder.ASCENDING
+  });
+  if (analyticUnits === null) {
+    return [];
+  }
+  return analyticUnits.map(createAnalyticUnitFromObject);
+}
+
+
+/**
+ * Creates and updates new unit.id
+ *
+ * @param unit to create
+ * @returns unit.id
+ */
+export async function create(unit: AnalyticUnit): Promise<AnalyticUnitId> {
+  let obj = unit.toObject();
+  return db.insertOne(obj);
+}
+
+export async function remove(id: AnalyticUnitId): Promise<void> {
+  // TODO: remove it`s segments
+  // TODO: remove it`s cache
+  await db.removeOne(id);
+}
+
+export async function update(id: AnalyticUnitId, unit: AnalyticUnit) {
+  const updateObj = {
+    name: unit.name,
+    labeledColor: unit.labeledColor,
+    deletedColor: unit.deletedColor,
+    visible: unit.visible
+  };
+
+  return db.updateOne(id, updateObj);
+}
+
+export async function setStatus(id: AnalyticUnitId, status: string, error?: string) {
+  return db.updateOne(id, { status, error });
+}
+
+export async function setDetectionTime(id: AnalyticUnitId, lastDetectionTime: number) {
+  return db.updateOne(id, { lastDetectionTime });
+}
+
+export async function setAlert(id: AnalyticUnitId, alert: boolean) {
+  return db.updateOne(id, { alert });
+}
+
+export async function setMetric(id: AnalyticUnitId, metric: Metric) {
+  return db.updateOne(id, { metric });
+}
