+++ server/src/models/analytics_task_model.ts
@@ -1,4 +1,4 @@
-import { AnalyticUnitId } from "./analytic_units/analytic_unit_model";
+import { AnalyticUnitId } from './analytic_units';
 
 import { uid } from "../utils/uid";
